Fix notifiying people multiple times for the same game (#10)
Sometimes people get notfied multiple times for the same game, see https://discord.com/channels/1275497005791051827/1336434968750198914/1340788242286645291 for example
I assume this happens due to the hash of the game changing midgame, due to a bookmove, that was not present prior, being added.
This occasionally happens in games (Im assuming due to instant moves being played and the GUI not handling it (the lack of UCI info?) "properly" and marking thos moves as book moves),
 for example https://ccrl.live/pgns/Zangdar_Gauntlet/20250216_2107_Nalwald_19_vs_Zangdar_3.04.pgn moves 161 black and following.

This is an attempt to fix this behaviour, by changing the filter when hashing the moves to a take_while. As far as I understood the documentation this should
now stop after the first non book move was found, which means this should not affect "normal" games, as bookmoves are expected to only be at the beginning of a game.

This also adds a test that checks that this case is handled correctly, this test is passed with the new version, however fails with the current version.

Note that I dont know any rust, I checked that this compiles, it might however break stuff that im not aware of, so its probably better to double check that this makes sense